[core] change std::array<uint16_t, 2> to mbgl::Size

+++ src/mbgl/sprite/sprite_parser.hpp
@@ -17,10 +17,10 @@ using SpriteImagePtr = std::shared_ptr<const SpriteImage>;
 
 // Extracts an individual image from a spritesheet from the given location.
 SpriteImagePtr createSpriteImage(const PremultipliedImage&,
-                                 uint16_t srcX,
-                                 uint16_t srcY,
-                                 uint16_t srcWidth,
-                                 uint16_t srcHeight,
+                                 uint32_t srcX,
+                                 uint32_t srcY,
+                                 uint32_t srcWidth,
+                                 uint32_t srcHeight,
                                  double ratio,
                                  bool sdf);
